Use dummy rate limiter in tests
In the tests calling the rate limiter has no effect. Compile time choose
the implementation module and delegate all the logic to it.
