+++ lib/sanbase/external_services/rate_limiting/behavior.ex
@@ -1,0 +1,9 @@
+defmodule Sanbase.ExternalServices.RateLimiting.Behavior do
+  @callback child_spec(name :: atom, options :: list()) :: %{
+              required(:id) => atom(),
+              required(:start) => {atom(), atom(), list()}
+            }
+
+  @callback wait(name :: atom()) :: :ok
+  @callback wait_until(name :: atom(), datetime :: DateTime.t()) :: :ok
+end
