+++ lib/sanbase/external_services/rate_limiting/test_server.ex
@@ -1,0 +1,33 @@
+defmodule Sanbase.ExternalServices.RateLimiting.TestServer do
+  use GenServer, restart: :permanent, shutdown: 5_000
+  @behaviour Sanbase.ExternalServices.RateLimiting.Behavior
+
+  require Logger
+
+  def child_spec(name, options \\ []) do
+    %{
+      id: name,
+      start: {
+        __MODULE__,
+        :start_link,
+        [Keyword.put(options, :name, name)]
+      }
+    }
+  end
+
+  def start_link(options) do
+    GenServer.start_link(__MODULE__, nil, name: Keyword.get(options, :name))
+  end
+
+  def init(_) do
+    {:ok, nil}
+  end
+
+  def wait(_) do
+    :ok
+  end
+
+  def wait_until(_, _) do
+    :ok
+  end
+end
