+++ lib/sanbase/scrapers.ex
@@ -1,4 +1,5 @@
 defmodule Sanbase.Application.Scrapers do
+  require Sanbase.Utils.Config, as: Config
   def init(), do: :ok
 
   @doc ~s"""
